requirements.txt
google-adk>=0.1.0
openai>=1.0.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pytz>=2024.1
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
alembic>=1.17.0
email-validator>=2.0.0
redis>=5.0.0
httpx>=0.24.0
# MCP (Model Context Protocol) dependencies
# Note: Using httpx for Notion API (already included above)
cryptography>=41.0.0  # For encrypting MCP credentials
nest-asyncio>=1.6.0  # For nested async event loops
# LiteLLM - Unified LLM Gateway
litellm>=1.50.0  # For unified LLM interface supporting 100+ providers

